quitado resources y mejorado developer
