+++ app/src/main/java/edu/iesam/gametracker/app/data/local/db/GameTrackerDataBase.kt
@@ -4,9 +4,8 @@ import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
 import edu.iesam.gametracker.app.data.local.db.converters.Converters
-import edu.iesam.gametracker.features.setting.data.local.DeveloperEntity
-import edu.iesam.gametracker.features.setting.data.local.ResourceEntity
-import edu.iesam.gametracker.features.setting.data.local.SettingDao
+import edu.iesam.gametracker.features.setting.data.developer.local.DeveloperEntity
+import edu.iesam.gametracker.features.setting.data.developer.local.DeveloperDao
 import edu.iesam.gametracker.features.videogames.data.local.db.FavoriteDao
 import edu.iesam.gametracker.features.videogames.data.local.db.FavoriteEntity
 import edu.iesam.gametracker.features.videogames.data.local.db.VideogamesDao
@@ -16,10 +15,9 @@ import edu.iesam.gametracker.features.videogames.data.local.db.VideogamesEntity
     entities = [
         VideogamesEntity::class,
         FavoriteEntity::class,
-        ResourceEntity::class,
         DeveloperEntity::class
     ],
-    version = 10,
+    version = 11,
     exportSchema = false
 )
 
@@ -27,5 +25,5 @@ import edu.iesam.gametracker.features.videogames.data.local.db.VideogamesEntity
 abstract class GameTrackerDataBase : RoomDatabase() {
     abstract fun videogamesDao(): VideogamesDao
     abstract fun favoriteDao(): FavoriteDao
-    abstract fun settingDao(): SettingDao
+    abstract fun settingDao(): DeveloperDao
 }
