+++ app/src/main/java/edu/iesam/gametracker/features/setting/data/developer/local/DbMappers.kt
@@ -1,0 +1,19 @@
+package edu.iesam.gametracker.features.setting.data.developer.local
+
+import edu.iesam.gametracker.features.setting.domain.developer.Developer
+
+fun Developer.toEntity(): DeveloperEntity {
+    return DeveloperEntity(
+        this.id,
+        this.name,
+        this.avatar
+    )
+}
+
+fun DeveloperEntity.toDomain(): Developer {
+    return Developer(
+        this.id,
+        this.name,
+        this.avatar
+    )
+}
