+++ app/src/main/java/edu/iesam/gametracker/features/setting/data/developer/local/DeveloperDao.kt
@@ -1,4 +1,4 @@
-package edu.iesam.gametracker.features.setting.data.local
+package edu.iesam.gametracker.features.setting.data.developer.local
 
 import androidx.room.Dao
 import androidx.room.Insert
@@ -6,13 +6,7 @@ import androidx.room.OnConflictStrategy
 import androidx.room.Query
 
 @Dao
-interface SettingDao {
-
-    @Query("SELECT * FROM $RESOURCE_TABLE")
-    suspend fun findAllResources(): List<ResourceEntity>
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun saveResources(vararg resources: ResourceEntity)
+interface DeveloperDao {
 
     @Query("SELECT * FROM $DEVELOPER_TABLE")
     suspend fun findAllDevelopers(): List<DeveloperEntity>
