+++ app/src/main/java/edu/iesam/gametracker/features/setting/data/developer/remote/DeveloperFirebaseMapper.kt
@@ -1,0 +1,12 @@
+package edu.iesam.gametracker.features.setting.data.developer.remote
+
+import edu.iesam.gametracker.features.setting.domain.developer.Developer
+
+
+fun DeveloperFirebaseModel.toModel(): Developer {
+    return Developer(
+        this.id,
+        this.name,
+        this.avatar
+    )
+}
