+++ app/src/main/java/edu/iesam/gametracker/features/setting/data/developer/remote/DeveloperFirebaseModel.kt
@@ -1,4 +1,4 @@
-package edu.iesam.gametracker.features.setting.data.remote
+package edu.iesam.gametracker.features.setting.data.developer.remote
 
 class DeveloperFirebaseModel(
     var id: Int = 0,
