--- app/src/main/java/edu/iesam/gametracker/features/setting/domain/Resource.kt
@@ -1,7 +1,0 @@
-package edu.iesam.gametracker.features.setting.domain
-
-data class Resource(
-    val id: Int,
-    val title: String,
-    val urlResource: String
-)
