--- app/src/main/java/edu/iesam/gametracker/features/setting/domain/SettingRepository.kt
@@ -1,7 +1,0 @@
-package edu.iesam.gametracker.features.setting.domain
-
-interface SettingRepository {
-    suspend fun getDevelopers(): Result<List<Developer>>
-
-    suspend fun getResources(): Result<List<Resource>>
-}
