+++ app/src/main/java/edu/iesam/gametracker/features/setting/domain/developer/DeveloperRepository.kt
@@ -1,0 +1,5 @@
+package edu.iesam.gametracker.features.setting.domain.developer
+
+interface DeveloperRepository {
+    suspend fun getDevelopers(): Result<List<Developer>>
+}
