+++ app/src/main/java/edu/iesam/gametracker/features/setting/domain/developer/GetDeveloperUseCase.kt
@@ -1,9 +1,9 @@
-package edu.iesam.gametracker.features.setting.domain
+package edu.iesam.gametracker.features.setting.domain.developer
 
 import org.koin.core.annotation.Single
 
 @Single
-class GetDeveloperUseCase(private val settingRepository: SettingRepository) {
+class GetDeveloperUseCase(private val settingRepository: DeveloperRepository) {
     suspend operator fun invoke(): Result<List<Developer>> {
         return settingRepository.getDevelopers()
     }
